Sync

+++ GADATA/GLUE/Tables/vultijd_data2.sql
@@ -3,7 +3,9 @@
     [controller] VARCHAR (50) NOT NULL,
     [Str1]       INT          NULL,
     [Str2]       INT          NULL,
-    [ID]         INT          IDENTITY (1, 1) NOT NULL,
+    [ID]         BIGINT       IDENTITY (1, 1) NOT NULL,
     CONSTRAINT [PK_vultijd_data2] PRIMARY KEY CLUSTERED ([ID] ASC)
 );
+
+
