+++ GADATA/GLUE/Tables/vultijden_viewers.sql
@@ -3,8 +3,10 @@
     [c_timestamp]   DATETIME NOT NULL,
     [Controller_id] INT      NOT NULL,
     [vultijd_ms]    INT      NOT NULL,
-    [couten]        INT      NULL,
+    [couten]        BIGINT   NULL,
     CONSTRAINT [PK_vultijden_viewers] PRIMARY KEY CLUSTERED ([ID] ASC),
     CONSTRAINT [FK_vultijden_viewers] FOREIGN KEY ([Controller_id]) REFERENCES [GLUE].[Controller] ([ID])
 );
+
+
